Add buf.ml
The idea is to have a more efficient and less error prone Cstruct.t.

Basically, we create Buf.t with 256 bytes, and the buffer automagically grow as
we add stuff. The nice thing is that it allows us to have this idiom:

     add_id SSH_MSG_DEBUG (create ()) |>
      add_bool always_display |>
      add_string message |>
      add_string lang

Before it would be something like

let a = encode_bool always_display in
let b = encode_string message in
let c = encode_string lang in

Cstruct.concat [a;b;c]

Each encode would be one alloc, then we copy all the data to each buffer, and
then later we recopy them while concatenating, that's ETOOMANYCOPIES.

The negative point is that encoding uses Buf.t while decoding uses Cstruct.t.