Implement Encode.encrypt.
Had to do some changes in kex.ml that I would not like to, in the future a
better abstraction will emerge.

I'm a bit puzzled why Nocrypto CTR modifies ctr/iv in place, and CBC doesn't, I
prefer much the CBC form since it is pure.

For now I'm emulating CBC interface in CTR, which means alocating a cstruct,
copying the old ctr and adding.