Fix encryption/decryption and add tests.
We can't just decrypt the whole input buffer, that ends up decrypting the mac as
well. We could keep track of the encrypted and unencrypted buffer but that
sucks.

So peek into the packet header and check if we have the whole message, this
wastes 4 bytes of extra encryption per message (or a block ?) but it's ok for
now.