Make Server.pop_msg return a Ssh.message and mock Ssh_version.
The previous code returned the buffer of a message since some Message handling
depends on the incoming buffer being unmodified. Essentialy, Ssh_kex_init uses
the incoming kex packet to derive the keys, so we can't affort to Decode/Encode
as the conversion may not identical (Decode x |> Encode is not x).

With this we can mock a Ssh_version on the input stage, so pop_msg can return a
Ssh_version message even though it's not a real message.

This also allows us to fix Case 5 of t_mpint, which shouldn't be there ffs.