Warp Kex.sign into Hostkey.sign, improve key signing & tests.
The signature of rsa keys will always be a SHA1 hash, so it makes no sense for
them to stay in Kex.

Hostkey.unsign was an error prone operation, change it to verify, which hides
the SHA1 hashing from the caller and it's way simpler.

Improve the signing tests, corrupt one byte at a time of the signature and check
that Hostkey.verify fails in all cases.

This corrupts byte 0, then restores byte 0, then corrupts byte 1 and so on, and
makes sure every one byte would fail.