Add option to ignore the next packet
The RFC 4253 states in section 7 that first_kex_packet_follows:

  If the other party's guess was wrong, and this field was TRUE,
  the next packet MUST be silently ignored,

In this change I add an option to the server record that will
indicate if the next package should be dropped silently.
And I add code, that drops the packet if this option is true.

Also I dropped the buf_of_message function which was not as useful
as the comment stated.