Client version is the whole string, with the comments.
Many many thanks to Konrad for helping me debug this, the error case was very
late, since the version string is used in key exchange to calculate H. So H
would be incorrect and the RSA sign/verification would fail.

To make things worse, only ubuntu used comments in the string, so every other
system worked fine, including compiling a vanilla openssh in the ubuntu machine.