Improve server main loop, allows us to send proper disconnects.
This introduces Ssh.action, which is the output of processing an output msg.
Every time you handle an input SSH message with Server.handle_msg, you are given
a list of Ssh.msg. Each msg in this list must be turned into an action by the
caller, which in turn is done with Server.action_of_msg.

An action can be sending data, disconnecting, or so on, priviously we had no way
to properly communicate the caller to send something _and_ disconnect, so we
were erroring out.