Rework a bit of channel_open.
There is no need to express the channel type both as a string and then the type
in channel_data, so do the same as we did in auth_method, where the ocaml type
expresses the string.

Since channel_open_sucess depends on context to express which channel data it
has, take it as a Cstruct, the caller has to convert with a function that shall
be provided, Frederick !

Code doesn't deal with direct-tcpip in all cases yet.