Handle SSH_MSG_CHANNEL_CLOSE and improve Channels a bit.
A channel is fully closed when both sides sent their CHANNEL_CLOSE. Since we
will always destroy the channel when it is fully closed, we must only express
the idea of a channel being "Open" or "Sent_close".

When we receive CLOSE and are at Open, we send our CLOSE and destroy the
channel.

There is still a lot of work pending, I still don't know how to build the API
for continuous channel data exchange, so focus should be on clearing the easy
bits for now.