Refactor Packet and keys.iv
Most operations of Packet dealt with passing keys up and down. This was a
decision since I considered it to be less error prone than passing cipher and
mac down, and then maybe forgetting to update keys somehow.

That proved itself to be wrong since IV is a property of the cipher, not keys.
When moving IV into the cipher, dealing with keys became ridiculous, so this
diff addresses some points:

* Move IV into cipher.
* Packet operations that don't require mac, now take a cipher as argument, and
IV is already there.
* Packet.encrypt and Packet.decrypt are the only functions that can reason on
keys, they just pass mac and keys down.
* Actually account TX bytes on encrypt, and RX bytes on decrypt. The count
represents bytes on the wire, so before decryption and after encryption.