Initial infrastructure for rekeying, disabled.
We account how many bytes we sent if this key, if it's over one GB,
Kex.should_rekey will be true.

Rekeying can be triggered even when not needed through Server.rekey, but, in
case we are already keying (or rekeying), rekeying becomes a nop.

Engine checks if it should rekey every time it produces Output.

Kex.should_rekey must be expanded to take into account how old our key is,
until we have a timing infrastructure, life is a bitch.